+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # CHANGELOG
 
+## 2.66.0
+
+### Minor Changes
+
+- [#375](https://github.com/labd/commercetools-node-mock/pull/375) [`5cbccde`](https://github.com/labd/commercetools-node-mock/commit/5cbccde6cd0c22a8cd7657e8224f22f6508af244) Thanks [@BramKaashoek](https://github.com/BramKaashoek)! - include custom type in Order Return Items
+
+## 2.65.1
+
+### Patch Changes
+
+- [#373](https://github.com/labd/commercetools-node-mock/pull/373) [`0729881`](https://github.com/labd/commercetools-node-mock/commit/0729881f84338984f2473cd695b40dd2e669289b) Thanks [@stephanbeek](https://github.com/stephanbeek)! - Added support for addDelivery order update action
+
 ## 2.65.0
 
 ### Minor Changes